Dirwatcher - long running program
Dirwatcher turned out to be more challenging then I thought it would be.  Keeping track of the lines that was read and ensuring that it was able to continue for that point on was difficult.

Although, this is my initial commit, it is the completed version of the program.  I'll return to it later to refactor it.